requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.0.6
djangorestframework==3.14.0
filelock==3.8.0
mysqlclient==2.1.1
Pillow==9.3.0
platformdirs==2.5.2
pytz==2022.4
sqlparse==0.4.2
tzdata==2022.1
python
